src/requirements.txt
SQLAlchemy
python-telegram-bot
psycopg2
stanza
seqlog
schedule
pytz
envparse